+++ wwr_ui/options.py
@@ -181,6 +181,10 @@ OPTIONS = OrderedDict([
     "If the chosen custom model comes with custom voice files, you can check this option to turn them off and simply use Link's normal voice instead."
   ),
   (
+    "disable_custom_player_items",
+    "If the chosen custom model comes with custom item models, you can check this option to turn them off and simply use Link's normal item models instead."
+  ),
+  (
     "custom_color_preset",
     "This allows you to select from preset color combinations chosen by the author of the selected player model."
   ),
@@ -215,6 +219,7 @@ NON_PERMALINK_OPTIONS = [
   "custom_player_model",
   "player_in_casual_clothes",
   "disable_custom_player_voice",
+  "disable_custom_player_items",
   "custom_color_preset",
   # Note: Options that affect music must be included in the permalink because music duration affects gameplay in some cases, like not being allowed to close the item get textbox until the item get jingle has finished playing.
 ]