+++ wwr_ui/randomizer_window.py
@@ -866,6 +866,12 @@ class WWRandomizerWindow(QMainWindow):
         self.ui.disable_custom_player_voice.show()
       else:
         self.ui.disable_custom_player_voice.hide()
+    
+    # Hide the custom items disable option for Link, but not any other models (since we don't know which have custom items).
+    if custom_model_name == "Link":
+      self.ui.disable_custom_player_items.hide()
+    else:
+      self.ui.disable_custom_player_items.show()
   
   def reload_colors(self, update_preview=True):
     preset_name = self.get_option_value("custom_color_preset")